requirements.txt
Django==5.0.1
Pillow==10.1.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
gunicorn==21.2.0
whitenoise==6.6.0
django-environ==0.11.2
